requirements.txt
# Core scientific stack
numpy==2.3.2
pandas==2.3.2
scipy==1.16.1

# Plotting & visualization
matplotlib==3.10.6
seaborn==0.13.2
pillow==11.3.0
fonttools==4.59.2
cycler==0.12.1
kiwisolver==1.4.9
contourpy==1.3.3
pyparsing==3.2.3

# Graph analysis & neuroimaging
networkx==3.5
bctpy==0.6.1
python-louvain==0.16
nibabel==5.3.2

# Utilities
packaging==25.0
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
typing-extensions==4.15.0
tzdata==2025.2
